requirements.txt
rich>=13.0.0
typer>=0.9.0
pyyaml>=6.0
watchdog>=3.0.0